rag_project/requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
tkinter
